backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography] 